2010-06-04  Ramon Novoa  <rnovoa@artica.es>
	* pandora_agent.spec: Fixed. Upgrading an RPM broke the installation.




git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@2858 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_agents/unix/pandora_agent.spec
@@ -88,6 +88,11 @@ chkconfig pandora_agent_daemon on
 
 %preun
 
+# Upgrading
+if [ "$1" = "1" ]; then
+        exit 0
+fi
+
 chkconfig --del pandora_agent_daemon 
 /etc/init.d/pandora_agent_daemon stop
 rm /etc/init.d/pandora_agent_daemon
